Fix MySQL connection issue in Docker setup

+++ db/migrate/20240923125823_remove_unnecessary_columns_from_users.rb
@@ -1,9 +1,9 @@
 class RemoveUnnecessaryColumnsFromUsers < ActiveRecord::Migration[6.0]
   def change
-    remove_column :users, :salt, :string
-    remove_column :users, :reset_password_token_expires_at, :datetime
-    remove_column :users, :remember_me_token, :string
-    remove_column :users, :remember_me_token_expires_at, :datetime
-    remove_column :users, :password_digest, :string
+    remove_column :users, :salt, :string if column_exists?(:users, :salt)
+    remove_column :users, :reset_password_token_expires_at, :datetime if column_exists?(:users, :reset_password_token_expires_at)
+    remove_column :users, :remember_me_token, :string if column_exists?(:users, :remember_me_token)
+    remove_column :users, :remember_me_token_expires_at, :datetime if column_exists?(:users, :remember_me_token_expires_at)
+    remove_column :users, :password_digest, :string if column_exists?(:users, :password_digest)
   end
 end
